feat: update property and task types to allow optional IDs; add latitude and longitude to property type; implement property management functions in supabase client
