+++ app/(pages)/(profile)/(houseOwner)/_layout.tsx
@@ -1,4 +1,7 @@
-import { Stack } from "expo-router";
+import React from 'react';
+import { router, Stack } from 'expo-router';
+import { AntDesign } from '@expo/vector-icons';
+import { TouchableOpacity } from 'react-native';
 
 export default function RootLayout() {
   return <RootLayoutNav />;
@@ -6,9 +9,24 @@ export default function RootLayout() {
 
 function RootLayoutNav() {
   return (
-    <Stack screenOptions={{ headerShown: false }}>
-      <Stack.Screen name="houseOwner" options={{ headerShown: false }} />
+    <Stack>
+      <Stack.Screen
+        name="createProperty"
+        options={{
+          headerTitle: 'Create Property',
+          headerTitleStyle: {
+            fontSize: 20,
+            fontWeight: 'bold',
+          },
+          headerLeft: () => (
+            <TouchableOpacity onPress={() => router.back()}>
+              <AntDesign name="left" size={24} color="black" />
+            </TouchableOpacity>
+          ),
+        }}
+      />
       <Stack.Screen name="expectedPricing" options={{ headerShown: false }} />
+      <Stack.Screen name="editProperty" options={{ headerShown: false }} />
     </Stack>
   );
 }
